Provide preference page to specify repository locations
This change introduces a 'Repository Locations' preference page in which
users can enter locations of repositries. To these repositories can then
be switched to with 'cr repo', where 'repo' is the before-last segment of
the repository directory. E.g. /path/to/foo/.git would be named 'foo'.
If the git console is opened for the first time, the first repository
is the current one.

Related issue: #4