Change eclipse target-4.4 to use Eclipse SDK
After deleting the workspace's bundle-pool location, AllPDETests did not run anymore
once the target bundles were reloaded. It seems that we were using the wrong SDK. Switching to
Eclipse SDK solved the problem. This also solved the 'duplicate manifest entry' warning,
which was the initial objective of my efforts.

The latter warning was caused by the source bundle of org.apache.felix.configadmin, which
contains duplicates indeed. However, our target does not reference that bundle, hence the
deletion of the bundle pool. I don't know how it got there initially and why the launch pay
attention to it as it is NOT part of the target and does NOT show up in the launch-config's
bundle list. Probably it was part of the initial target and maybe the launch check's the
validity of the complete pool, but that's just wild speculation.

I keep an eye on it...