Adds a CONTRIBUTING.md
The text of this file will be shown when users open pull requests. It also ensures that contributers agree
to the license terms of the projects. This is useful because it ensures that all code is realy compatible
with the EPL and e.g. can be contributed back to Eclipse without legal problems.