Adjust dependency version ranges to work with Eclipse Mars
With Eclipse 4.5 (Mars) EGit and JGit changed their version to 4.x to
allow for breaking API changes.
Luckily, Gonsole is not affected from these API changes. Changing the
version range for dependencies on EGit and JGit to [3.7,5.0) allows to
continue to work with older versions of Eclipse/EGit and also support
Mars.

See #53 Gonsole EGit integration not installable on Eclipse Mars