Adapt to recent changes in the EGit Activator
The EGit `Activator` used to provide access to the `RepositoryUtil`
that ist used to list the configured repositories in EGit. In recent
versions of EGit access was moved to `RepositoryUtil::getInstance`.

Use reflection to try to get `RepositoryUtil`, first the _old way_,
then the _new way_.

Fix: https://github.com/rherrmann/gonsole/issues/82